+++ django_stripe/apps.py
@@ -1,0 +1,24 @@
+import stripe
+
+import django_stripe
+from django.apps import AppConfig as DjangoAppConfig
+
+
+class AppConfig(DjangoAppConfig):
+
+    name = "django_stripe"
+    verbose_name = "Stripe"
+
+    def ready(self) -> None:
+        from . import settings
+
+        # Stripe library setup
+        stripe.api_key = settings.SECRET_KEY
+        stripe.api_version = settings.API_VERSION
+
+        # Identify the plugin to Stripe
+        stripe.set_app_info(
+            "django-stripe-lite",
+            version=django_stripe.__version__,
+            url="https://github.com/yunojuno/django-stripe-lite",
+        )
